+++ packages/@expo/cli/src/export/fork-bundleAsync.ts
@@ -14,6 +14,7 @@ import chalk from 'chalk';
 import Metro from 'metro';
 import { Terminal } from 'metro-core';
 
+import { WebSupportProjectPrerequisite } from '../start/doctor/web/WebSupportProjectPrerequisite';
 import { MetroTerminalReporter } from '../start/server/metro/MetroTerminalReporter';
 import { withMetroMultiPlatform } from '../start/server/metro/withMetroMultiPlatform';
 import { getPlatformBundlers } from '../start/server/platformBundlers';
@@ -106,7 +107,14 @@ export async function bundleAsync(
 
   const { exp } = getConfig(projectRoot, { skipSDKVersionRequirement: true });
   let config = await ExpoMetroConfig.loadAsync(projectRoot, { reporter, ...options });
-  config = withMetroMultiPlatform(projectRoot, config, getPlatformBundlers(exp));
+
+  const bundlerPlatforms = getPlatformBundlers(exp);
+
+  if (bundlerPlatforms.web === 'metro') {
+    await new WebSupportProjectPrerequisite(projectRoot).assertAsync();
+  }
+
+  config = withMetroMultiPlatform(projectRoot, config, bundlerPlatforms);
 
   const metroServer = await metro.runMetro(config, {
     watch: false,